# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dexe Investing Api",
        "title": "Dexe Investing Api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/pool-transfers/deposits/{wallet}": {
            "get": {
                "description": "Get Deposits By Wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PoolTransfers"
                ],
                "summary": "Get Deposits By Wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet Address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.PoolTransfer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/pool-transfers/withdrawals/{wallet}": {
            "get": {
                "description": "Get Withdrawals By Wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PoolTransfers"
                ],
                "summary": "Get Withdrawals By Wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet Address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.PoolTransfer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/pools": {
            "get": {
                "description": "Get All Trader Pools",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pools"
                ],
                "summary": "Get All Trader Pools",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Pool"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/pools/{wallet}": {
            "get": {
                "description": "Get Pools By Creator Wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pools"
                ],
                "summary": "Get Pools By Creator Wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pool creator wallet",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Pool"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/trader/{traderWallet}": {
            "get": {
                "description": "Get Trader Info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trader"
                ],
                "summary": "Get Trader Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trader wallet address",
                        "name": "traderWallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service_routes.TraderInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/trades/{traderPool}": {
            "get": {
                "description": "Get Trades By PoolAdr",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trades"
                ],
                "summary": "Get Trades By PoolAdr",
                "parameters": [
                    {
                        "type": "string",
                        "description": "traderPool Address",
                        "name": "traderPool",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/service_routes.responseTrade"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Pool": {
            "type": "object",
            "properties": {
                "basicTokenAdr": {
                    "type": "string"
                },
                "blockNumber": {
                    "description": "blockNumber bigint,",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "creatorAdr": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "dexeCommissionDen": {
                    "type": "integer"
                },
                "dexeCommissionNum": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "investorCommissionDen": {
                    "type": "integer"
                },
                "investorCommissionNum": {
                    "type": "integer"
                },
                "investorRestricted": {
                    "type": "boolean"
                },
                "isActualOn": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "poolAdr": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "totalSupply": {
                    "type": "string"
                },
                "traderCommissionDen": {
                    "type": "integer"
                },
                "traderCommissionNum": {
                    "type": "integer"
                },
                "tx": {
                    "description": "tx character varying(255),",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.PoolTransfer": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "blockNumber": {
                    "description": "blockNumber bigint,",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "poolAdr": {
                    "type": "string"
                },
                "tx": {
                    "description": "tx character varying(255),",
                    "type": "string"
                },
                "type": {
                    "description": "enum('deposit', 'withdraw')",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "wallet": {
                    "type": "string"
                }
            }
        },
...
        "response.E": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.S": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "datetime": {
                    "type": "string"
                }
            }
        },
...
        "service_routes.TraderInfoResponse": {
            "type": "object",
            "properties": {
                "copiers": {
                    "type": "number"
                },
                "fund": {
                    "type": "integer"
                }
            }
        },
        "service_routes.responseTrade": {
            "type": "object",
            "properties": {
                "amountIn": {
                    "type": "string"
                },
                "amountOutMin": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "path": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "traderPool": {
                    "type": "string"
                },
                "tx": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}